fixed lesson UI
