+++ src/screens/courseManager/addLesson/AddLesson.jsx
@@ -9,15 +9,14 @@ import { addVideo } from './addVideo'
 import { addText } from './addText'
 import { addImage } from './addImage'
 
+import AddContent from './AddContent'
+import LessonMaterial from './LessonMaterial'
+import LessonSaveModal from './LessonSaveModal'
 import AddTestModal from '../../../components/addTestModal/AddTestModal'
 import BackButton from '../../../components/backButton/BackButton'
-import DragDrop from '../../../components/dragDrop/DragDrop'
 import NewsCreateModal from '../../../components/newsCreateModal/NewsCreateModal'
 import DashboardLayout from '../../../layout/dashboardLayout/DashboardLayout'
-import { ErrorText } from '../../../components/formUI/FormUI'
-import Video from '../../../components/videoPlayer/Video'
-import Image from '../../../components/lessonImage/Image'
-import Text from './Text'
+
 import './AddLesson.scss'
 
 const AddLesson = () => {
@@ -38,6 +37,8 @@ const AddLesson = () => {
   const [material, setMaterial] = useState(null)
 
   const { register, errors, handleSubmit } = useForm()
+
+  console.log(lessonData)
 
   const submitLessonForm = ({ title }) => {
     const coverImg = lessonCover
@@ -125,108 +126,6 @@ const AddLesson = () => {
         <LessonSaveModal onClick={handleSubmit(submitLessonForm)} />
       </DashboardLayout>
     </>
-  )
-}
-
-const AddContent = ({
-  setVideoModal,
-  setImageModal,
-  setTextModal,
-  setTestModal,
-  register,
-  errors,
-  setLessonCover,
-  lessonData
-}) => {
-  return (
-    <div className='admin-lesson-create-container'>
-      <input
-        type='text'
-        placeholder='Write title here'
-        name='title'
-        ref={register({
-          required: {
-            value: true,
-            message: 'You must enter lesson title'
-          }
-        })}
-      />
-      <ErrorText
-        className='errorMsg'
-        message={errors.title && errors.title.message}
-      />
-      <DragDrop onChange={(img) => setLessonCover(img)} />
-      {lessonData &&
-        lessonData.map((vid, index) => (
-          <div key={index}>
-            <Video
-              title={vid.videoTitle}
-              description={vid.videoDescription}
-              url={vid.videoLink || vid.videoResource?.preview}
-              thumbnail={vid.videoCover?.preview}
-            />
-            <Image src={vid.lessonImg?.preview} desc={vid.photoDescription} />
-            <Text heading={vid.textHeading} desc={vid.textDescription} />
-          </div>
-        ))}
-      <div className='admin-lesson-create-btn-wrapper'>
-        <button className='secondary-btn' onClick={() => setVideoModal(true)}>
-          <img src='/img/video-outline.svg' alt='video icon' />{' '}
-          <span>Add video</span>
-        </button>
-        <button className='secondary-btn' onClick={() => setImageModal(true)}>
-          <img src='/img/image-outline.svg' alt='image_icon' />
-          <span>Add image</span>
-        </button>
-        <button className='secondary-btn' onClick={() => setTextModal(true)}>
-          <img src='/img/text-outline.svg' alt='text icon' />{' '}
-          <span>Add text</span>
-        </button>
-        <button className='secondary-btn' onClick={() => setTestModal(true)}>
-          <img src='/img/test-outline.svg' alt='test icon' />{' '}
-          <span>Add test</span>
-        </button>
-      </div>
-    </div>
-  )
-}
-
-const LessonMaterial = ({ material, setMaterial }) => {
-  return (
-    <div className='admin-lesson-materials-container'>
-      <h1>Materials</h1>
-      {material ? (
-        <p>{material.name}</p>
-      ) : (
-        <p>You dont have any materials in lesson Add it for your users.</p>
-      )}
-      <DragDrop
-        fileType='application/pdf'
-        className='secondary-btn addMaterial'
-        text='Add Materials'
-        onChange={(mat) => setMaterial(mat)}
-      />
-    </div>
-  )
-}
-
-const LessonSaveModal = ({ onClick }) => {
-  return (
-    <div className='save-lesson-modal'>
-      <h4>Do you want to save lesson?</h4>
-      <div>
-        <button className='secondary-btn' id='lesson-cancel-btn'>
-          Cancel
-        </button>
-        <button
-          className='primary-btn secondary-btn'
-          id='lesson-save-btn'
-          onClick={onClick}
-        >
-          Save lesson
-        </button>
-      </div>
-    </div>
   )
 }
